Go back to generating firing rates on the fly
Previously Experiment.generate_spike_rates would be run once to create a
KDE of the session which is then used by align_trials. The KDE was
replaced with a gaussian convolution which is must faster so performance
is no longer an issue, and the Experiment-level method made things
awkward on the cache side when trying different kernel sizes. Instead,
sigma=X can be passed to align_trials(..., 'spike_rate') and used for
filtering.