Rewrite TDMS->AVI conversion to avoid memory errors
This replaces the TDMS->AVI ioutils function with another that doesn't
load all of the data into memory at once, instead loading and saving one
frame at a time.

Videos are changed back to AVI as the error turned out to be due to
forcing the output frame rate to an unexpected value, which causing
frame drops or duplication by ffmpeg.